[nmpc][py]: move sim scripts to the outside of package; refactor sim_fake_sensor.py to fake_sensor.py, refactor sim_fir_differentiator.py to fir_differentiator.py
