+++ aerial_robot_control/scripts/nmpc/nmpc_tilt_mt/tilt_qd/fake_sensor.py
