+++ aerial_robot_control/scripts/nmpc/nmpc_tilt_mt/tilt_qd/tilt_qd_servo_dist_imp.py
@@ -4,7 +4,7 @@ import numpy as np
 import casadi as ca
 
 from nmpc_tilt_mt.tilt_qd.qd_nmpc_base import QDNMPCBase
-from nmpc_tilt_mt.tilt_qd.sim_fake_sensor import FakeSensor
+from nmpc_tilt_mt.tilt_qd.fake_sensor import FakeSensor
 from nmpc_tilt_mt.tilt_qd import phys_param_beetle_omni as phys_omni
