+++ aerial_robot_control/scripts/nmpc/sim_nmpc_impedance.py
@@ -6,7 +6,7 @@ import argparse
 
 from nmpc_tilt_mt.nmpc_viz import Visualizer
 
-from nmpc_tilt_mt.tilt_qd.sim_fir_differentiator import FIRDifferentiator
+from nmpc_tilt_mt.tilt_qd.fir_differentiator import FIRDifferentiator
 
 from nmpc_tilt_mt.tilt_qd.tilt_qd_servo_thrust_dist_imp import NMPCTiltQdServoThrustImpedance
 from nmpc_tilt_mt.tilt_qd.tilt_qd_servo_thrust_dist import NMPCTiltQdServoThrustDist
